handshakes are working as expected
